Change to unsigned version

+++ headers/CPU.h
@@ -12,9 +12,9 @@ public:
     CPU() : PC(0) {};
     void step();
     void loadInstruction(uint16_t addr, uint16_t instruction); // Load an instruction into memory at a specific address
-    uint16_t getRegisterValue(uint8_t reg) const {return rs.read(reg);} 
+    uint16_t getRegisterValue(uint8_t reg) const {return (int16_t)rs.read(reg);} 
     uint16_t getCycles() const {return cycles.get();}
-    void setRegisterValue(uint8_t reg, uint16_t value) { rs.write(reg, value);}
+    void setRegisterValue(uint8_t reg, int16_t value) { rs.write(reg, uint16_t(value));}
     void setPC(uint16_t value) { PC = value; }
     uint16_t getPC() const {return PC;}
