+++ headers/Instruction.h
@@ -7,8 +7,8 @@ struct Instruction {
     uint8_t rd;
     uint8_t rs1;
     uint8_t rs2;
-    uint8_t addr1; 
-    int8_t imm;
+    int16_t addr1; 
+    int16_t imm;
 
 };
